+++ packages/tricorder/src/Tricorder/Session/CabalFile.hs
@@ -1,7 +1,9 @@
 module Tricorder.Session.CabalFile
     ( CabalFile (..)
     , inputCabalFiles
-    , discoverCabalFiles
+    , discoverPackages
+    , discoverCabalPackages
+    , discoverStackPackages
     )
 where
 
@@ -14,16 +16,22 @@ import Data.Traversable (for)
 import Distribution.Fields (Field (..), FieldLine (..), Name (..), readFields)
 import Distribution.PackageDescription.Parsec (parseGenericPackageDescriptionMaybe)
 import Distribution.Types.GenericPackageDescription (GenericPackageDescription)
+import Effectful.Exception (throwIO)
 import Effectful.Reader.Static (Reader, ask)
 import System.FilePath (normalise, takeExtension, (</>))
 import System.FilePath.Glob (compile)
+import System.IO.Error (userError)
 
 import Atelier.Effects.Env qualified as Env
+import Atelier.Effects.FileSystem qualified as FileSystem
 import Atelier.Effects.Log qualified as Log
 import Data.ByteString.Char8 qualified as BC
 import Data.Text qualified as T
 
 import Tricorder.Runtime (ProjectRoot (..))
+import Tricorder.Session.StackYaml (StackYaml)
+
+import Tricorder.Session.StackYaml qualified as StackYaml
 
 
 data CabalFile = CabalFile
@@ -39,40 +47,55 @@ inputCabalFiles
        , Glob :> es
        , Log :> es
        , Reader ProjectRoot :> es
+       , StackYaml :> es
        )
     => Eff (Input [CabalFile] : es) a -> Eff es a
 inputCabalFiles = runInputEff do
-    projectFilePaths <- discoverCabalFiles
-    (faileds, packageDescriptions) <-
-        partitionEithers <$> for projectFilePaths \p -> do
-            contents <- readFileBs p
-            case parseGenericPackageDescriptionMaybe contents of
-                Nothing -> pure $ Left p
-                Just gpd -> pure $ Right $ CabalFile p gpd
-    unless (null faileds) do
-        Log.warn
-            $ "Failed to parse .cabal files for the following packages: "
-                <> T.intercalate ", " (toText <$> faileds)
-    pure $ packageDescriptions
+    packageRes <- discoverPackages
+    case packageRes of
+        Left err -> throwIO $ userError $ toString err
+        Right projectFilePaths -> do
+            (faileds, packageDescriptions) <-
+                partitionEithers <$> for projectFilePaths \p -> do
+                    contents <- readFileBs p
+                    case parseGenericPackageDescriptionMaybe contents of
+                        Nothing -> pure $ Left p
+                        Just gpd -> pure $ Right $ CabalFile p gpd
+            unless (null faileds) do
+                Log.warn
+                    $ "Failed to parse .cabal files for the following packages: "
+                        <> T.intercalate ", " (toText <$> faileds)
+            pure $ packageDescriptions
+
+
+discoverPackages
+    :: (Env :> es, FileSystem :> es, Glob :> es, Reader ProjectRoot :> es, StackYaml :> es)
+    => Eff es (Either Text [FilePath])
+discoverPackages = do
+    ProjectRoot projectRoot <- ask
+    hasStackYaml <- FileSystem.doesFileExist $ projectRoot </> "stack.yaml"
+    if hasStackYaml
+        then discoverStackPackages
+        else discoverCabalPackages
 
 
 -- | Discovers `.cabal` files in all locations and formats Cabal itself
 -- supports.
-discoverCabalFiles
+discoverCabalPackages
     :: ( Env :> es
        , FileSystem :> es
        , Glob :> es
        , Reader ProjectRoot :> es
        )
-    => Eff es [FilePath]
-discoverCabalFiles = do
+    => Eff es (Either Text [FilePath])
+discoverCabalPackages = do
     ProjectRoot projectRoot <- ask
     homeCabalFiles <- maybe [] (one . (</> ".cabal/config")) <$> Env.lookupEnv "HOME"
     let projectFilePaths = projectCabalFiles projectRoot <> homeCabalFiles
     projectFiles <- filterM doesFileExist projectFilePaths
     if null projectFiles
         then
-            cabalFilesIn projectRoot
+            Right <$> cabalFilesIn projectRoot
         else do
             packages <- fmap (find (not . null))
                 $ for projectFiles \projectFile -> do
@@ -82,8 +105,8 @@ discoverCabalFiles = do
                             (cabalFilesForEntry projectRoot)
                             (projectPackageEntries contents)
             case packages of
-                Nothing -> cabalFilesIn projectRoot
-                Just pkgs -> pure pkgs
+                Nothing -> Right <$> cabalFilesIn projectRoot
+                Just pkgs -> pure $ Right pkgs
   where
     projectCabalFiles projectRoot =
         (projectRoot </>) <$> ["cabal.project.local", "cabal.project.freeze", "cabal.project"]
@@ -98,6 +121,17 @@ discoverCabalFiles = do
         resolveMatch path
             | isCabalFile path = pure [path]
             | otherwise = cabalFilesIn path
+
+
+discoverStackPackages
+    :: (Reader ProjectRoot :> es, StackYaml :> es)
+    => Eff es (Either Text [FilePath])
+discoverStackPackages = do
+    ProjectRoot projectRoot <- ask
+    mProject <- StackYaml.readProject
+    case mProject of
+        Left err -> pure $ Left $ "Failed to read project file: " <> err
+        Right project -> pure $ Right $ normalise . (projectRoot </>) <$> project.packages
 
 
 -- | List the @.cabal@ files directly inside a directory.